Add MVC models

+++ Models.Global/Data/Category.cs
@@ -10,6 +10,6 @@ namespace Models.Global.Data
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public float? BudgetLimit { get; set; }
+        public double? BudgetLimit { get; set; }
     }
 }
